move each session test to a separate file to avoid false results. fixes #58
