+++ src/test/java/com/cognifide/aemrules/checks/SessionShouldBeLoggedOutTest.java
@@ -20,14 +20,45 @@
 package com.cognifide.aemrules.checks;
 
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
 
+import java.util.Arrays;
+import java.util.Collection;
+
+@RunWith(Parameterized.class)
 public class SessionShouldBeLoggedOutTest extends AbstractBaseTest {
+
+	// format is filename, failure expected
+	@Parameterized.Parameters
+	public static Collection<Object[]> data() {
+		return Arrays.asList(new Object[][] {
+				{ "src/test/files/checks/SessionLogoutOne.java", false},
+				{ "src/test/files/checks/SessionLogoutTwo.java", false},
+				{ "src/test/files/checks/SessionLogoutThree.java", false},
+				{ "src/test/files/checks/SessionLogoutFour.java", true},
+				{ "src/test/files/checks/SessionLogoutFive.java", true},
+				{ "src/test/files/checks/SessionLogoutSix.java", true},
+				{ "src/test/files/checks/SessionLogoutSeven.java", true},
+				{ "src/test/files/checks/SessionLogoutEight.java", false}
+		});
+	}
+
+	private boolean expectFailure;
+
+	public SessionShouldBeLoggedOutTest(Object fn, Object expectFailure) {
+		filename = (String) fn;
+		this.expectFailure = ((Boolean) expectFailure).booleanValue();
+	}
 
 	@Test
 	public void checkInjectorNotClosedInFinallyBlock() {
 		check = new SessionShouldBeLoggedOut();
-		filename = "src/test/files/checks/SessionLogout.java";
-		verify();
+		if (expectFailure) {
+			verify();
+		} else {
+			verifyNoIssues();
+		}
 	}
 
 }
